healthd: Reinitialize mChargerNames for every battery update
Booting up the device without usb, the kernel sets the usb power supply
type as UNKNOWN. The type of usb power supply changes at run-time as
various chargers are plugged in/out. However, healthd initilizes the
charger list only at bootup. Change it such that it checks for charger
type changes with every battery or usb uevent.

While at it, the kernel may have a power supply type which is not known
to healthd. This is perfectly fine. Update healthd to not print a
warning.

Change-Id: I2ec9f9a420ca61814d43c316b418ce94de3691bc

Former-commit-id: 08a453898adf2f5bfb594396d6d915d68ebb67b8
Former-commit-id: f5eebbe10af6f7cdeffc928dda25e1b70eef498e
Change-Id: I8fe57f5f8114c3827e025fb7cdb7503991665507

Former-commit-id: 806d87599f8c2c6da97ff4409c946dc0f225ead1

Former-commit-id: ce701561a1a3710fc8f5d57f609160306735ad78
Signed-off-by: S I P U N <sipunkumar85@gmail.com>